Decoupled commonlib from calculators
