+++ openquake/commands/run.py
@@ -25,10 +25,10 @@ import io
 
 from openquake.baselib import performance, general
 from openquake.risklib import valid
-from openquake.commonlib import sap, readinput, datastore, oqvalidation
+from openquake.commonlib import sap, readinput, datastore, oqvalidation, views
 from openquake.commonlib.concurrent_futures_process_mpatch import (
     concurrent_futures_process_monkeypatch)
-from openquake.calculators import base, views
+from openquake.calculators import base
 CT = oqvalidation.OqParam.concurrent_tasks.default
 
 calc_path = None  # set only when the flag --slowest is given
