+++ openquake/engine/bin/openquake_cli.py
@@ -59,8 +59,7 @@ from openquake.engine import engine, logs
 from openquake.engine.tools.make_html_report import make_report
 from openquake.engine.export import core
 from openquake.risklib import valid
-from openquake.commonlib import datastore
-from openquake.calculators import views
+from openquake.commonlib import datastore, views
 
 from openquake.commonlib.concurrent_futures_process_mpatch import (
     concurrent_futures_process_monkeypatch)
