+++ openquake/engine/engine.py
@@ -28,8 +28,8 @@ from openquake.baselib.performance import Monitor
 from openquake.risklib import valid
 from openquake.commonlib import parallel, readinput
 from openquake.commonlib.oqvalidation import OqParam
-from openquake.commonlib import export, datastore
-from openquake.calculators import base, views
+from openquake.commonlib import export, datastore, views
+from openquake.calculators import base
 from openquake.engine import logs, config
 
 TERMINATE = valid.boolean(
